requirements.txt
Flask
Flask-Migrate
flask-restx
Flask-SQLAlchemy
Werkzeug
psycopg2-binary
pyjwt
datetime
uuid
argon2-cffi
uwsgi
# Flask==2.1.2
# Flask-Migrate==3.1.0
# flask-restx==0.5.1
# Flask-SQLAlchemy==2.5.1
# Werkzeug==2.1.2
# psycopg2-binary
# pyjwt==2.0.0
# datetime
# uuid
# argon2-cffi
# uwsgi